add CLI command to run the script without having to define env

+++ core/reststarter.py
@@ -22,7 +22,7 @@ import sys
 import time
 import psutil
 from common import consts
-from utils import restutils
+from utils import restutils, oshelper
 from core.baseappsrv import BaseAppServer
 
 
@@ -35,46 +35,50 @@ class RESTServerStarter(BaseAppServer):
         super(RESTServerStarter, self).do_configure()
         
     def do_start(self):
-        pid_file = consts.DEFAULT_SCRIPT_PATH + '/data/temp/irest.pid'
-        os.makedirs(os.path.dirname(pid_file), exist_ok=True)
-        run_args = [
-            'gunicorn',
-            '-w', '4',
-            '-k', 'uvicorn.workers.UvicornWorker',
-            '-b', '0.0.0.0:8080',
-            '-n', 'centric-rest-ibridge',
-            '-p', str(pid_file),
-            '--error-logfile', consts.DEFAULT_RESTAPI_LOG_FILE
-        ]
-        run_args += ["""irest:create_app()"""]
-        gunicorn_master_proc = None
+        # only enable this on production mode in linux
+        if self.is_production_mode() and oshelper.is_linux():
+            pid_file = consts.DEFAULT_SCRIPT_PATH + '/data/temp/irest.pid'
+            os.makedirs(os.path.dirname(pid_file), exist_ok=True)
+            run_args = [
+                'gunicorn',
+                '-w', '4',
+                '-k', 'uvicorn.workers.UvicornWorker',
+                '-b', '0.0.0.0:8080',
+                '-n', 'centric-rest-ibridge',
+                '-p', str(pid_file),
+                '--error-logfile', consts.DEFAULT_RESTAPI_LOG_FILE
+            ]
+            run_args += ["""irest:create_app()"""]
+            gunicorn_master_proc = None
 
-        def kill_proc(dummy_signum, dummy_frame):
-            gunicorn_master_proc.terminate()
-            gunicorn_master_proc.wait()
-            sys.exit(0)
+            def kill_proc(dummy_signum, dummy_frame):
+                gunicorn_master_proc.terminate()
+                gunicorn_master_proc.wait()
+                sys.exit(0)
 
-        gunicorn_master_proc = subprocess.Popen(run_args)
+            gunicorn_master_proc = subprocess.Popen(run_args)
 
-        signal.signal(signal.SIGINT, kill_proc)
-        signal.signal(signal.SIGTERM, kill_proc)
+            signal.signal(signal.SIGINT, kill_proc)
+            signal.signal(signal.SIGTERM, kill_proc)
 
         restutils.set_stopped(False)
         super(RESTServerStarter, self).do_start()
 
     def do_stop(self):
-        pid_file = consts.DEFAULT_SCRIPT_PATH + '/data/temp/irest.pid'
-        while True:
-            try:
-                with open(pid_file) as f:
-                    gunicorn_master_proc_pid = int(f.read())
-                    break
-            except IOError:
-                logging.debug("Waiting for gunicorn's pid file to be created.")
-                time.sleep(0.1)
+        # only enable this on production mode in linux
+        if self.is_production_mode() and oshelper.is_linux():
+            pid_file = consts.DEFAULT_SCRIPT_PATH + '/data/temp/irest.pid'
+            while True:
+                try:
+                    with open(pid_file) as f:
+                        gunicorn_master_proc_pid = int(f.read())
+                        break
+                except IOError:
+                    logging.debug("Waiting for gunicorn's pid file to be created.")
+                    time.sleep(0.1)
 
-        gunicorn_master_proc = psutil.Process(gunicorn_master_proc_pid)
-        gunicorn_master_proc.terminate()
+            gunicorn_master_proc = psutil.Process(gunicorn_master_proc_pid)
+            gunicorn_master_proc.terminate()
         restutils.set_stopped(True)
         super(RESTServerStarter, self).do_stop()
